+++ NEWS.md
@@ -2,6 +2,7 @@ The latest release of the SAMtool package is available on [CRAN](https://CRAN.R-
 
 ## SAMtool 1.2.5
 - `SCA_RWM` can accept multiple years to the `refyear` argument, e.g., `expression(1:Data@Year)`. The model will calculate reference points (MSY, unfished values, and steepness) using the mean M during the specified years.
+- Fix bug in interim MPs to return `NA` in `Rec@TAC` when multiple assessments do not converge.
 
 ## SAMtool 1.2.4
 - Update `Shortcut` indexing to align year of assessment with projection. An MP using the `Perfect` assessment and `HCR_MSY` annually will produce F = FMSY in the OM.